customer-service/requirements.txt
django==4.2
djangorestframework==3.14
mysqlclient==2.2.0
requests==2.31
django-cors-headers==4.3
